Always display help and errors assuming name "pow"
When pow is installed via npm, the way you call pow is kinda lost
and pow believes it's called via `pow.py` or `pow.exe` and it uses
those names in help and error messages. This is counter-intuitive
for the user, therefore, I'm hereby making pow dumber and letting it
believe it is always called as `pow` (as it should realistically).